Separate python/intepreter api and extension api
