+++ src/extension.web.ts
@@ -39,7 +39,7 @@ import {
     window,
     workspace
 } from 'vscode';
-import { buildApi, IExtensionApi } from './platform/api';
+import { buildApi, IExtensionApi } from './webviews/extension-side/api/api';
 import { IApplicationEnvironment, ICommandManager } from './platform/common/application/types';
 import { traceError } from './platform/logging';
 import {
