+++ src/test/common.ts
@@ -4,7 +4,7 @@
 'use strict';
 
 import { NotebookDocument, Uri, Event } from 'vscode';
-import { IExtensionApi } from '../platform/api';
+import { IExtensionApi } from '../webviews/extension-side/api/api';
 import { IDisposable } from '../platform/common/types';
 import { IServiceContainer, IServiceManager } from '../platform/ioc/types';
