+++ src/test/datascience/notebook/exportFull.vscode.test.ts
@@ -35,9 +35,9 @@ import { BufferDecoder } from '../../../platform/common/process/decoder.node';
 import { INbConvertInterpreterDependencyChecker, INotebookImporter } from '../../../kernels/jupyter/types';
 import { JupyterImporter } from '../../../webviews/extension-side/import-export/jupyterImporter.node';
 import { IInterpreterService } from '../../../platform/interpreter/contracts';
-import { PythonEnvironment } from '../../../platform/api/extension';
 import { CodeSnippets, Identifiers } from '../../../platform/common/constants';
 import { noop } from '../../../platform/common/utils/misc';
+import { PythonEnvironment } from '../../../platform/pythonEnvironments/info';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-require-imports
 const expectedPromptMessageSuffix = `requires ${ProductNames.get(Product.ipykernel)!} to be installed.`;
