+++ src/webviews/extension-side/api/kernelApi.ts
@@ -3,19 +3,19 @@
 
 import { injectable, inject } from 'inversify';
 import { Disposable, Event, EventEmitter, Uri } from 'vscode';
-import { KernelConnectionWrapper } from './kernelConnectionWrapper';
+import { KernelConnectionWrapper } from '../../../platform/api/kernelConnectionWrapper';
 import {
     IKernelProvider,
     IKernel,
     KernelConnectionMetadata as IKernelKernelConnectionMetadata
-} from '../../kernels/types';
-import { INotebookControllerManager } from '../../notebooks/types';
-import { disposeAllDisposables } from '../common/helpers';
-import { traceInfo } from '../logging';
-import { IDisposable, IDisposableRegistry, IExtensions } from '../common/types';
-import { PromiseChain } from '../common/utils/async';
-import { IKernelSocket as ExtensionKernelSocket } from '../../kernels/types';
-import { sendTelemetryEvent } from '../../telemetry';
+} from '../../../kernels/types';
+import { INotebookControllerManager } from '../../../notebooks/types';
+import { disposeAllDisposables } from '../../../platform/common/helpers';
+import { traceInfo } from '../../../platform/logging';
+import { IDisposable, IDisposableRegistry, IExtensions } from '../../../platform/common/types';
+import { PromiseChain } from '../../../platform/common/utils/async';
+import { IKernelSocket as ExtensionKernelSocket } from '../../../kernels/types';
+import { sendTelemetryEvent } from '../../../telemetry';
 import { ApiAccessService } from './apiAccessService';
 import {
     ActiveKernel,
@@ -25,11 +25,11 @@ import {
     KernelConnectionMetadata,
     WebSocketData
 } from './extension';
-import { JupyterNotebookView, Telemetry } from '../common/constants';
-import { KernelConnector } from '../../notebooks/controllers/kernelConnector';
-import { DisplayOptions } from '../../kernels/displayOptions';
-import { IServiceContainer } from '../ioc/types';
-import { IExportedKernelServiceFactory } from './types';
+import { JupyterNotebookView, Telemetry } from '../../../platform/common/constants';
+import { KernelConnector } from '../../../notebooks/controllers/kernelConnector';
+import { DisplayOptions } from '../../../kernels/displayOptions';
+import { IServiceContainer } from '../../../platform/ioc/types';
+import { IExportedKernelServiceFactory } from './api';
 
 @injectable()
 export class JupyterKernelServiceFactory implements IExportedKernelServiceFactory {
