+++ src/webviews/extension-side/serviceRegistry.node.ts
@@ -41,6 +41,9 @@ import { ActiveEditorContextService } from './activeEditorContext';
 import { AmlComputeContext } from './amlContext.node';
 import { IImportTracker, ImportTracker } from './importTracker.node';
 import { GlobalActivation } from './globalActivation';
+import { JupyterKernelServiceFactory } from './api/kernelApi';
+import { IExportedKernelServiceFactory } from './api/api';
+import { ApiAccessService } from './api/apiAccessService';
 
 export function registerTypes(serviceManager: IServiceManager, _isDevMode: boolean) {
     serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, GlobalActivation);
@@ -105,4 +108,11 @@ export function registerTypes(serviceManager: IServiceManager, _isDevMode: boole
     );
 
     serviceManager.addSingletonInstance<IExtensionSideRenderer>(IExtensionSideRenderer, new ExtensionSideRenderer());
+
+    // API
+    serviceManager.addSingleton<IExportedKernelServiceFactory>(
+        IExportedKernelServiceFactory,
+        JupyterKernelServiceFactory
+    );
+    serviceManager.addSingleton<ApiAccessService>(ApiAccessService, ApiAccessService);
 }
